Add files via upload
 Project Overview

This pharmacy project is a comprehensive management system designed to streamline various processes and tasks involved in running a pharmacy. It offers a range of features to facilitate efficient inventory management, Admin handling, and billing operations. Below is a brief demonstration of some key functionalities:

Inventory Management

Add Item: The system allows you to add new items to the inventory. For instance, you can add "Aspirin" with a quantity of 50, an expiry date of 2024-06-30, and a price of $2.50.
Update Item: Let's say you want to update the quantity of "Aspirin" to 100. You can easily modify the item's details in the inventory.
Remove Item: If an item is no longer available, such as "Ibuprofen," you can remove it from the inventory.

Billing System

Calculate Bill: Suppose a customer named "Jane Smith" purchases 2 units of "Aspirin." The billing system will calculate the total cost based on the item price and quantity.
Generate Itemized Bill: The system can generate an itemized bill for "Jane Smith," displaying the details of the purchased items, quantities, individual prices, and the total amount due.

For more visualization, I've shared a video.